+++ src/ts/main/index.ts
@@ -1,13 +1,13 @@
 import { app, BrowserWindow, ipcMain, dialog , nativeImage, globalShortcut } from "electron";
 import { MainFolder, WindowSettings } from "./main-folder.js";
-import { downloadYoutubeSong } from "./download-song.js";
+import { downloadYoutubeSong, getYoutubeSongSrc } from "./song.js";
 import * as path from "path";
 
 class Index {
     private window: BrowserWindow | null = null;
     private readonly mainFolder: MainFolder;
     private thumbarButtons: Electron.ThumbarButton[] = [];
-    private readonly dev: boolean = true;
+    private readonly dev: boolean = false;
 
     constructor() {
         this.mainFolder = new MainFolder();
@@ -190,6 +190,14 @@ class Index {
             }
 
             return await downloadYoutubeSong(this.window!, videoID);
+        });
+
+        ipcMain.handle("youtube-getSongSrc", async (_event, videoID: string) => {
+            if (this.window == null) {
+                throw new Error("Can't get youtube song src: Window is null.");
+            }
+
+            return await getYoutubeSongSrc(this.window!, videoID);
         });
     }