+++ changelog.md
@@ -1,5 +1,6 @@
-## HEKO-3.1.1
-
+## HEKO-3.2.0
+#### ADDED
+- add songs from youtube
 
 ## HEKO-3.1.0
 #### ADDED